+++ CHANGELOG.md
@@ -10,12 +10,21 @@
 ## [Unreleased][]
 ### Added
 - '09_statistics.js' new example showing a timeline per artist/# tracks.
+- 'timeline' graph type used specifically to display 3-D data along the new 'multi' graph variable (see below). Is a modified version of the 'bars' type.
 - 'bAltLabels' axis variable for alternative drawing. For bars, displays the text in vertical (so it doesn't get cut on small width). For doughnut and pie charts, draws a line from the serie to the label.
 - 'bPopupBackground' configuration variable to display an overlay while loading Async data. Default behavior is false, i.e. only the text and animation is displayed.
 - 'bProfile' configuration variable to enable profiling logging on console.
 - 'bSlicePerKey' configuration variable to force slicing manipulation by total number of x asis keys, instead of values per series. For ex. 2 series may have different X values, resulting on an X axis longer than expected since slicing gets N values per serie, not the values associated to the first N X-keys shared by all series. This is now the default behavior.
 - 'multi' graph variable to display aggregated data per X-value, i.e. effectively displaying 3-dimensional data. For ex. number of tracks (Y) per Artist (Z) per Year (X), where artists will be grouped per year. This type of data is better displayed on bar graphs. Data should be passed with this schema: [[[{x1, y11, z11}, ...],[{x2, y21, z21}, ...], ...]], where a single serie contains multiple X-planes, each one being an array of points with same X-value. The framework will automatically manipulate the data to display the X-points of every X-plane array on different series. In case the X-plane arrays don't have the same number of points, output series may have different length, which is now handled by the 'bSlicePerKey' config above.
+- Built-in 'natural', 'invert', 'string natural', 'string invert', 'random' (Fisher-Yates algorithm) sorting methods, by setting dataManipulation.sort to those labels.
+- 'dataManipulation.sort' function may now be a method present on the array prototype, like: ```dataManipulation.sort = Array.prototype.shuffle;``` instead of a compare function  used within .sort().
+- 'dataManipulation.sort' function may be added as a pair [function, [args]], where the function is a method on array prototype and args is passed as array[function](...args). This may is usually used along Schwartzian transform for sorting; in fact the method is built-in and can be called with ```dataManipulation.sort = ['Schwartzian transform', [(point) => processPoint(point)]];``` or ```dataManipulation.sort = [Array.prototype.schwartzianSort, [(point) => processPoint(point)]];```, being equivalent. The second argument for the sorting method is the compare function, provided by default as natural sorting for numbers. Change as required for strings (for. ex. using localeCompare()).
+- Scroll, settings and display settings buttons have been added; enabled by setting ```buttons = {xScroll: true , settings: true, display: true}```. Functionality is defined by ```callbacks = {settings: {/* onLbtnUp, onRbtnUp */}, display: {/* onLbtnUp, onRbtnUp */}}```. Scroll functionality is internally handled.
+- Callback have been added, defined by ```{point: {/* onLbtnUp, onRbtnUp */}, focus: {/* onMouseWwheel, onRbtnUp */}, settings: {/* onLbtnUp, onRbtnUp */}, display: {/* onLbtnUp, onRbtnUp */}}```. 'onMouseWwheel' functionality (while panel is on focus) is internally handled by default, unless explicitly set to null or replaced.
+- Zoom functionality using the mouse wheel, which slices the data range to be smaller/bigger.
+- Scroll functionality using mouse dragging while clicking, which slices the data to the left/right (not changing the range size).
 ### Changed
+- Adjusted mouse cursor over specific elements. 
 - Improved chart menus with all new additions and code cleanup.
 - Tooltip now also shows percentages on pie and doughnut modes.
 - Background color can now be set to null to use it as overlay.
@@ -28,6 +37,7 @@
 - Improved data checking for bad inputs.
 - Y axis ticks were sometimes not properly set to respect the margins or panel size.
 - Fixed some setting not being set in some cases while using the menu.
+- Multiple minor UI fixes.
 
 ## [0.3.1] - 2023-08-24
 ### Added